Move infer out of `middle::typeck` and into just `middle`.
