+++ src/librustc/middle/infer/higher_ranked/doc.rs
