+++ src/librustc/middle/traits/fulfill.rs
@@ -10,7 +10,7 @@
 
 use middle::mem_categorization::Typer;
 use middle::ty;
-use middle::typeck::infer::InferCtxt;
+use middle::infer::InferCtxt;
 use std::collections::HashSet;
 use std::rc::Rc;
 use util::ppaux::Repr;
