+++ src/librustc/middle/traits/select.rs
@@ -30,8 +30,8 @@ use middle::fast_reject;
 use middle::mem_categorization::Typer;
 use middle::subst::{Subst, Substs, VecPerParamSpace};
 use middle::ty::{mod, Ty};
-use middle::typeck::infer;
-use middle::typeck::infer::{InferCtxt, TypeSkolemizer};
+use middle::infer;
+use middle::infer::{InferCtxt, TypeSkolemizer};
 use middle::ty_fold::TypeFoldable;
 use std::cell::RefCell;
 use std::collections::hash_map::HashMap;
