+++ src/librustc/middle/typeck/check/_match.rs
@@ -14,7 +14,7 @@ use middle::subst::{Subst, Substs};
 use middle::ty::{mod, Ty};
 use middle::typeck::check::{check_expr, check_expr_has_type, demand, FnCtxt};
 use middle::typeck::check::{instantiate_path, structurally_resolved_type, valid_range_bounds};
-use middle::typeck::infer::{mod, resolve};
+use middle::infer::{mod, resolve};
 use middle::typeck::require_same_types;
 use util::nodemap::FnvHashMap;
