+++ src/librustc/middle/typeck/check/method/probe.rs
@@ -21,8 +21,8 @@ use middle::ty::{MethodObject};
 use middle::ty_fold::HigherRankedFoldable;
 use middle::typeck::check;
 use middle::typeck::check::{FnCtxt, NoPreference};
-use middle::typeck::infer;
-use middle::typeck::infer::InferCtxt;
+use middle::infer;
+use middle::infer::InferCtxt;
 use syntax::ast;
 use syntax::codemap::{Span, DUMMY_SP};
 use std::collections::HashSet;
