+++ src/librustc/middle/typeck/check/regionck.rs
@@ -124,9 +124,9 @@ use middle::typeck::astconv::AstConv;
 use middle::typeck::check::FnCtxt;
 use middle::typeck::check::regionmanip;
 use middle::typeck::check::vtable;
-use middle::typeck::infer::resolve_and_force_all_but_regions;
-use middle::typeck::infer::resolve_type;
-use middle::typeck::infer;
+use middle::infer::resolve_and_force_all_but_regions;
+use middle::infer::resolve_type;
+use middle::infer;
 use middle::pat_util;
 use util::nodemap::{DefIdMap, NodeMap, FnvHashMap};
 use util::ppaux::{ty_to_string, Repr};
