+++ src/librustc/middle/typeck/check/writeback.rs
@@ -19,9 +19,9 @@ use middle::ty::{mod, Ty, MethodCall, MethodCallee};
 use middle::ty_fold::{TypeFolder,TypeFoldable};
 use middle::typeck::astconv::AstConv;
 use middle::typeck::check::FnCtxt;
-use middle::typeck::infer::{force_all, resolve_all, resolve_region};
-use middle::typeck::infer::resolve_type;
-use middle::typeck::infer;
+use middle::infer::{force_all, resolve_all, resolve_region};
+use middle::infer::resolve_type;
+use middle::infer;
 use middle::typeck::write_substs_to_tcx;
 use middle::typeck::write_ty_to_tcx;
 use util::ppaux::Repr;
