+++ src/librustc/middle/typeck/coherence/mod.rs
@@ -32,9 +32,9 @@ use middle::ty::type_is_ty_var;
 use middle::subst::Subst;
 use middle::ty;
 use middle::typeck::CrateCtxt;
-use middle::typeck::infer::combine::Combine;
-use middle::typeck::infer::InferCtxt;
-use middle::typeck::infer::{new_infer_ctxt, resolve_ivar, resolve_type};
+use middle::infer::combine::Combine;
+use middle::infer::InferCtxt;
+use middle::infer::{new_infer_ctxt, resolve_ivar, resolve_type};
 use std::collections::{HashSet};
 use std::cell::RefCell;
 use std::rc::Rc;
