+++ src/librustc_trans/test.rs
@@ -21,10 +21,10 @@ use middle::stability;
 use middle::subst;
 use middle::subst::Subst;
 use middle::ty::{mod, Ty};
-use middle::typeck::infer::combine::Combine;
-use middle::typeck::infer;
-use middle::typeck::infer::lub::Lub;
-use middle::typeck::infer::glb::Glb;
+use middle::infer::combine::Combine;
+use middle::infer;
+use middle::infer::lub::Lub;
+use middle::infer::glb::Glb;
 use session::{mod,config};
 use syntax::{abi, ast, ast_map, ast_util};
 use syntax::codemap;
